separate gui editor events and move side-bar type
